+++ CHANGELOG.md
@@ -21,6 +21,7 @@ project adheres to
 
 * Lots of color handling.
   - Spec changes for traditional css colors (PR #198).
+  - Made all color channels f64 instead of Rational (PR #199).
 * Updated sass-spec test suite to 2024-09-13.